requirements.txt
appnope==0.1.4
asttokens==3.0.0
audioop-lts==0.2.2
audioread==3.1.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
comm==0.2.3
debugpy==1.8.17
decorator==5.2.1
essentia==2.1b6.dev1389
executing==2.2.1
idna==3.11
ipykernel==7.1.0
ipython==9.7.0
ipython-pygments-lexers==1.1.1
jedi==0.19.2
joblib==1.5.2
jupyter-client==8.6.3
jupyter-core==5.9.1
lazy-loader==0.4
librosa==0.11.0
llvmlite==0.45.1
matplotlib-inline==0.2.1
msgpack==1.1.2
mutagen==1.47.0
nest-asyncio==1.6.0
numba==0.62.1
numpy==2.3.4
packaging==25.0
pandas==2.3.3
parso==0.8.5
pexpect==4.9.0
platformdirs==4.5.0
pooch==1.8.2
prompt-toolkit==3.0.52
psutil==7.1.3
psycopg2-binary==2.9.11
ptyprocess==0.7.0
pure-eval==0.2.3
py4j==0.10.9.9
pyarrow==22.0.0
pycparser==2.23
pygments==2.19.2
pyspark==4.0.1
python-dateutil==2.9.0.post0
pytz==2025.2
pyyaml==6.0.3
pyzmq==27.1.0
requests==2.32.5
scikit-learn==1.7.2
scipy==1.16.3
six==1.17.0
soundfile==0.13.1
soxr==1.0.0
sqlalchemy==2.0.44
stack-data==0.6.3
standard-aifc==3.13.0
standard-chunk==3.13.0
standard-sunau==3.13.0
threadpoolctl==3.6.0
tornado==6.5.2
traitlets==5.14.3
typing-extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
wcwidth==0.2.14
